Extract jsonapi-serializer in to a separate package
