+++ packages/@orbit/integration-tests/Brocfile.js
@@ -14,6 +14,7 @@ let buildOptions = {
     '@orbit/indexeddb',
     '@orbit/indexeddb-bucket',
     '@orbit/jsonapi',
+    '@orbit/jsonapi-serializer',
     '@orbit/local-storage',
     '@orbit/local-storage-bucket',
     '@orbit/record-cache',
@@ -34,6 +35,7 @@ if (process.env.BROCCOLI_ENV === 'tests') {
     packageDist('@orbit/indexeddb'),
     packageDist('@orbit/indexeddb-bucket'),
     packageDist('@orbit/jsonapi'),
+    packageDist('@orbit/jsonapi-serializer'),
     packageDist('@orbit/local-storage'),
     packageDist('@orbit/local-storage-bucket'),
     packageDist('@orbit/serializers'),
