+++ packages/@orbit/jsonapi-serializer/.npmignore
@@ -1,0 +1,12 @@
+docs
+node_modules
+test
+tests
+tmp
+*.log
+*.DS_Store
+*.editorconfig
+*.gitignore
+*.travis.yml
+Brocfile.js
+testem.json
