+++ packages/@orbit/jsonapi-serializer/Brocfile.js
@@ -2,17 +2,13 @@
 
 const build = require('@glimmer/build');
 const packageDist = require('@glimmer/build/lib/package-dist');
-const funnel = require('broccoli-funnel');
-const path = require('path');
 
 let buildOptions = {
   external: [
     '@orbit/utils',
     '@orbit/core',
     '@orbit/data',
-    '@orbit/serializers',
-    '@orbit/jsonapi-serializer',
-    'sinon'
+    '@orbit/serializers'
   ]
 };
 
@@ -21,10 +17,7 @@ if (process.env.BROCCOLI_ENV === 'tests') {
     packageDist('@orbit/utils'),
     packageDist('@orbit/core'),
     packageDist('@orbit/data'),
-    packageDist('@orbit/serializers'),
-    packageDist('@orbit/jsonapi-serializer'),
-    funnel(path.join(require.resolve('sinon'), '../../pkg'), { include: ['sinon.js'] }),
-    funnel(path.join(require.resolve('whatwg-fetch'), '../'), { include: ['fetch.js'] })
+    packageDist('@orbit/serializers')
   ];
 }
