+++ packages/@orbit/jsonapi-serializer/package.json
@@ -1,9 +1,10 @@
 {
-  "name": "@orbit/jsonapi",
-  "version": "0.16.0-beta.4",
-  "description": "JSON:API support for Orbit.",
+  "name": "@orbit/jsonapi-serializer",
+  "version": "0.16.0-beta.1",
+  "description": "JSON:API serializer for Orbit.",
   "contributors": [
-    "Dan Gebhardt <dan@cerebris.com>"
+    "Dan Gebhardt <dan@cerebris.com>",
+    "Paul Chavard <github@paul.chavard.net>"
   ],
   "keywords": [
     "orbit",
@@ -29,13 +30,9 @@
     "@orbit/core": "^0.16.0-beta.1",
     "@orbit/data": "^0.16.0-beta.1",
     "@orbit/serializers": "^0.16.0-beta.1",
-    "@orbit/jsonapi-serializer": "^0.16.0-beta.1",
     "@orbit/utils": "^0.16.0-beta.1"
   },
   "devDependencies": {
-    "@glimmer/build": "^0.10.1",
-    "@types/sinon": "7.0.3",
-    "sinon": "^7.2.2",
-    "whatwg-fetch": "^2.0.3"
+    "@glimmer/build": "^0.10.1"
   }
 }
