+++ packages/@orbit/jsonapi-serializer/src/index.ts
@@ -1,0 +1,3 @@
+export { default, JSONAPISerializerSettings, DeserializeOptions } from './jsonapi-serializer';
+export * from './record-document';
+export * from './resource-document';
