+++ packages/@orbit/jsonapi-serializer/test/.eslintrc.js
@@ -1,0 +1,15 @@
+module.exports = {
+  globals: {
+    module: true,
+    test: true,
+    ok: true,
+    strictEqual: true,
+    equal: true,
+    throws: true,
+    notEqual: true,
+    deepEqual: true,
+    expect: true,
+    QUnit: true,
+    notStrictEqual: true
+  }
+}
