+++ packages/@orbit/jsonapi-serializer/testem.json
@@ -1,0 +1,17 @@
+{
+  "framework": "qunit",
+  "test_page": "tests/index.html",
+  "on_start": "npm run build:tests",
+  "on_change": "npm run build:tests",
+  "src_files": ["src/**/*", "test/**/*"],
+  "launch_in_ci": ["Chrome"],
+  "browser_args": {
+    "mode": "ci",
+    "Chrome": [
+      "--headless",
+      "--disable-gpu",
+      "--remote-debugging-port=9222",
+      "--no-sandbox"
+    ]
+  }
+}
