+++ packages/@orbit/jsonapi/src/resource-document.ts
@@ -1,40 +1,7 @@
-import { Dict } from '@orbit/utils';
-import { Link } from '@orbit/data';
+import { ResourceDocument } from '@orbit/jsonapi-serializer';
 
-export interface ResourceIdentity {
-  id: string;
-  type: string;
-}
-
-export interface ResourceHasOneRelationship {
-  data?: ResourceIdentity | null;
-  meta?: Dict<any>;
-  links?: Dict<Link>;
-}
-
-export interface ResourceHasManyRelationship {
-  data?: ResourceIdentity[];
-  meta?: Dict<any>;
-  links?: Dict<Link>;
-}
-
-export type ResourceRelationship = ResourceHasOneRelationship | ResourceHasManyRelationship;
-
-export interface Resource {
-  id?: string;
-  type: string;
-  attributes?: Dict<any>;
-  relationships?: Dict<ResourceRelationship>;
-  meta?: Dict<any>;
-  links?: Dict<Link>;
-}
-
-export interface ResourceDocument {
-  data: Resource | Resource[] | ResourceIdentity | ResourceIdentity[];
-  included?: Resource[];
-  meta?: Dict<any>;
-  links?: Dict<Link>;
-}
+export { ResourceIdentity, ResourceHasOneRelationship, ResourceHasManyRelationship, ResourceRelationship, Resource } from '@orbit/jsonapi-serializer';
+export { ResourceDocument };
 
 /**
  * @deprecated
